requirements.txt
azure-functions
slack_webhook
